lambda-functions/requirements.txt
kafka_python==2.0.1
Faker==9.7.1
text_unidecode==1.3
urllib3==1.26.6
cfnresponse==1.1.2
typing_extensions==4.0.1